# admin/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Tabio POS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /outlets/{outletId}/orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: outletId
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdersResponse"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create order
      parameters:
        - name: outletId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /outlets/{outletId}/orders/{orderId}:
    get:
      operationId: getOrder
      tags: [orders]
      parameters:
        - name: outletId
          in: path
          required: true
          schema:
            type: integer
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    put:
      operationId: updateOrder
      tags: [orders]
      parameters:
        - name: outletId
          in: path
          required: true
          schema:
            type: integer
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /outlets/{outletId}/orders/{orderId}/bill:
    post:
      operationId: generateBill
      tags: [orders]
      summary: Generate bill/invoice
      parameters:
        - name: outletId
          in: path
          required: true
          schema:
            type: integer
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateBillRequest"
      responses:
        "200":
          description: Bill generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
  /outlets/{outletId}/orders/{orderId}/kot:
    post:
      operationId: generateKot
      tags: [kot]
      summary: Generate KOT
      parameters:
        - name: outletId
          in: path
          required: true
          schema:
            type: integer
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: KOT generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KOT"
...
  /outlets/{outletId}/aggregators/zomato/orders:
    get:
      operationId: getZomatoOrders
      tags: [aggregators]
      summary: Get Zomato orders (mock)
      parameters:
        - name: outletId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Zomato orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AggregatorOrder"
...
  /outlets/{outletId}/aggregators/swiggy/orders:
    get:
      operationId: getSwiggyOrders
      tags: [aggregators]
      parameters:
        - name: outletId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Swiggy orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AggregatorOrder"
...
components:
  schemas:
...
    Outlet:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
        gstNumber:
          type: string
        fssaiNumber:
          type: string
        currency:
          type: string
        timezone:
          type: string
        logoUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, name, address, phone, email, currency, timezone, isActive, createdAt]
...
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        menuItemId:
          type: integer
        menuItemName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        variantId:
          type: integer
          nullable: true
        variantName:
          type: string
          nullable: true
        addonIds:
          type: array
          items:
            type: integer
        addonNames:
          type: array
          items:
            type: string
        addonTotal:
          type: number
        totalPrice:
          type: number
        notes:
          type: string
          nullable: true
        kotStatus:
          type: string
          enum: [pending, preparing, ready, served]
      required: [id, menuItemId, menuItemName, quantity, unitPrice, addonIds, addonNames, addonTotal, totalPrice, kotStatus]
    Order:
      type: object
      properties:
        id:
          type: integer
        outletId:
          type: integer
        orderNumber:
          type: string
        type:
          type: string
          enum: [dine-in, takeaway, delivery, zomato, swiggy]
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, served, billed, paid, cancelled]
        tableId:
          type: integer
          nullable: true
        tableName:
          type: string
          nullable: true
        customerId:
          type: integer
          nullable: true
        customerName:
          type: string
          nullable: true
        customerPhone:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        discountAmount:
          type: number
        discountType:
          type: string
          nullable: true
        discountValue:
          type: number
          nullable: true
        taxAmount:
          type: number
        totalAmount:
          type: number
        paymentMethod:
          type: string
          nullable: true
          enum: [cash, card, upi, wallet, split, null]
        paymentStatus:
          type: string
          enum: [pending, partial, paid, refunded]
        notes:
          type: string
          nullable: true
        aggregatorOrderId:
          type: string
          nullable: true
        aggregatorPlatform:
          type: string
          nullable: true
        staffId:
          type: integer
          nullable: true
        staffName:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, outletId, orderNumber, type, status, items, subtotal, discountAmount, taxAmount, totalAmount, paymentStatus, createdAt, updatedAt]
    OrdersResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required: [orders, total, limit, offset]
    CreateOrderRequest:
      type: object
      properties:
        type:
          type: string
          enum: [dine-in, takeaway, delivery, zomato, swiggy]
        tableId:
          type: integer
          nullable: true
        customerId:
          type: integer
          nullable: true
        customerName:
          type: string
          nullable: true
        customerPhone:
          type: string
          nullable: true
        items:
          type: array
          items:
            type: object
            properties:
              menuItemId:
                type: integer
              quantity:
                type: integer
              variantId:
                type: integer
                nullable: true
              addonIds:
                type: array
                items:
                  type: integer
              notes:
                type: string
                nullable: true
            required: [menuItemId, quantity]
        notes:
          type: string
          nullable: true
        staffId:
          type: integer
          nullable: true
      required: [type, items]
    UpdateOrderRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, served, billed, paid, cancelled]
        items:
          type: array
          items:
            type: object
            properties:
              menuItemId:
                type: integer
              quantity:
                type: integer
              variantId:
                type: integer
                nullable: true
              addonIds:
                type: array
                items:
                  type: integer
              notes:
                type: string
                nullable: true
            required: [menuItemId, quantity]
        discountType:
          type: string
          nullable: true
        discountValue:
          type: number
          nullable: true
        paymentMethod:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        tableId:
          type: integer
          nullable: true
    GenerateBillRequest:
      type: object
      properties:
        paymentMethod:
          type: string
          enum: [cash, card, upi, wallet, split]
        splitPayments:
          type: array
          nullable: true
          items:
            type: object
            properties:
              method:
                type: string
              amount:
                type: number
            required: [method, amount]
        discountType:
          type: string
          nullable: true
        discountValue:
          type: number
          nullable: true
        customerId:
          type: integer
          nullable: true
        loyaltyPointsRedeemed:
          type: integer
          nullable: true
      required: [paymentMethod]
    Bill:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        billNumber:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        discountAmount:
          type: number
        taxAmount:
          type: number
        totalAmount:
          type: number
        paymentMethod:
          type: string
        splitPayments:
          type: array
          nullable: true
          items:
            type: object
            properties:
              method:
                type: string
              amount:
                type: number
            required: [method, amount]
        loyaltyPointsEarned:
          type: integer
        loyaltyPointsRedeemed:
          type: integer
        createdAt:
          type: string
        outlet:
          $ref: "#/components/schemas/Outlet"
      required: [id, orderId, billNumber, items, subtotal, discountAmount, taxAmount, totalAmount, paymentMethod, loyaltyPointsEarned, loyaltyPointsRedeemed, createdAt, outlet]
    KOT:
      type: object
      properties:
        id:
          type: integer
        outletId:
          type: integer
        orderId:
          type: integer
        orderNumber:
          type: string
        kotNumber:
          type: string
        tableId:
          type: integer
          nullable: true
        tableName:
          type: string
          nullable: true
        orderType:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        status:
          type: string
          enum: [pending, preparing, ready, served]
        station:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, outletId, orderId, orderNumber, kotNumber, orderType, items, status, station, createdAt, updatedAt]
...
    AggregatorOrder:
      type: object
      properties:
        id:
          type: string
        platform:
          type: string
          enum: [zomato, swiggy]
        status:
          type: string
          enum: [new, accepted, preparing, ready, picked, delivered, cancelled]
        customerName:
          type: string
        customerPhone:
          type: string
        deliveryAddress:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              quantity:
                type: integer
              price:
                type: number
            required: [name, quantity, price]
        subtotal:
          type: number
        totalAmount:
          type: number
        estimatedDelivery:
          type: string
        createdAt:
          type: string
      required: [id, platform, status, customerName, customerPhone, deliveryAddress, items, subtotal, totalAmount, estimatedDelivery, createdAt]
...