# admin/lib/api-zod/src/generated/types/createOrderRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tabio POS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderRequestItemsItem } from "./createOrderRequestItemsItem";
import type { CreateOrderRequestType } from "./createOrderRequestType";

export interface CreateOrderRequest {
  type: CreateOrderRequestType;
  tableId?: number | null;
  customerId?: number | null;
  customerName?: string | null;
  customerPhone?: string | null;
  items: CreateOrderRequestItemsItem[];
  notes?: string | null;
  staffId?: number | null;
}